# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Market APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/purchase": {
            "post": {
                "tags": [
                    "Purchase"
                ],
                "summary": "Create a new purchase",
                "description": "Create a new purchase",
                "operationId": "createPurchase",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "provider_id",
                                    "storage_id",
                                    "products"
                                ],
                                "properties": {
                                    "provider_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "storage_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "products": {
                                        "description": "Product details",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity",
                                                "unit_price"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": "2"
                                                },
                                                "unit_price": {
                                                    "type": "numeric",
                                                    "example": "1000.00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error on validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Incorrect data."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error in server."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/refund": {
            "post": {
                "tags": [
                    "Refund"
                ],
                "summary": "Create a new refund",
                "description": "Create a new refund",
                "operationId": "createRefund",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "batch_id",
                                    "storage_id",
                                    "refund_type",
                                    "products"
                                ],
                                "properties": {
                                    "batch_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "storage_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "refund_type": {
                                        "description": "The type of refund, either 'purchase' or 'sale'",
                                        "type": "string",
                                        "enum": [
                                            "purchase",
                                            "sale"
                                        ],
                                        "example": "purchase"
                                    },
                                    "products": {
                                        "description": "Product details",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity",
                                                "unit_price"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": "2"
                                                },
                                                "unit_price": {
                                                    "type": "numeric",
                                                    "example": "1000.00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error on validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Incorrect data."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error in server."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}